migrate to MySQL and implement users profile control

+++ app.py
@@ -1,11 +1,12 @@
 from flask import Flask, jsonify, request
 from models.user import User
 from database import db
+import bcrypt
 from flask_login import LoginManager, login_user, current_user, logout_user, login_required
 
 app = Flask(__name__)
 app.config["SECRET_KEY"] = "password"
-app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///database.db"
+app.config["SQLALCHEMY_DATABASE_URI"] = "mysql+pymysql://root:administrator@127.0.0.1:3306/flask-crud"
 
 login_manager = LoginManager()
 db.init_app(app)
@@ -24,7 +25,7 @@ def login():
     password = data.get("password")
     if username and password:
         user = User.query.filter_by(username=username).first()
-        if user and user.password == password:
+        if user and bcrypt.checkpw(str.encode(password), str.encode(user.password)):
             login_user(user)
             return jsonify({"message": "Autenticação realizada."}), 200
         
@@ -41,9 +42,11 @@ def create_user():
     data = request.json
     username = data.get("username")
     password = data.get("password")
+    role = data.get("role", "user")
 
     if username and password:
-        user = User(username=username, password=password)
+        hashed_password = bcrypt.hashpw(str.encode(password), bcrypt.gensalt())
+        user = User(username=username, password=hashed_password, role=role)
         db.session.add(user)
         db.session.commit()
         return jsonify({"message": "Usuário cadastrado."}), 201
@@ -55,7 +58,7 @@ def create_user():
 def read_users():
     users = User.query.all()
     if users:
-        return jsonify({"users": [{"username": user.username} for user in users]}), 200
+        return jsonify({"users": [{"username": user.username, "role": user.role} for user in users]}), 200
         
     return jsonify({"message":"Nenhum usuário cadastrado"}), 404
 
@@ -64,7 +67,7 @@ def read_users():
 def read_user(id_user):
     user = User.query.get(id_user)
     if user:
-        return {"username": user.username}
+        return {"username": user.username, "role": user.role}
         
     return jsonify({"message":"Usuário não encotrado"}), 404
 
@@ -73,10 +76,13 @@ def read_user(id_user):
 def update_user(id_user):
     data = request.json
     user = User.query.get(id_user)
+    if id_user != current_user.id and current_user.role == "user":
+        return jsonify({"message": "Operação não permitida"}), 403
+
     if user and data.get("password"):
         user.password = data.get("password")
         db.session.commit()
-        return jsonify({"message": f"Usuário id:{id_user} atualizado. Senha {user.password}"})
+        return jsonify({"message": f"Senha alterada para usuário id {id_user}"})
     
 
     return jsonify({"message":"Usuário não encontrado"}), 404
@@ -85,8 +91,13 @@ def update_user(id_user):
 @login_required
 def delete_user(id_user):
     user = User.query.get(id_user)
+
+    if current_user.role != "admin":
+        return jsonify({"message": "Operação não permitida"}), 403
+    
     if id_user == current_user.id:
         return jsonify({"message":"Não permitido excluir este usuário"}), 403
+    
     if user:
         db.session.delete(user)
         db.session.commit()
